backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-chess

